chore: Split Rust files up into separate files
